Update Register-PSResourceRepository.md - Example 4 (#225)
* Update Register-PSResourceRepository.md - Example 4

Provide a little more information around example 4. From my reading here it wasn't clear anywhere that the secret must be stored as a PSCredential object and expected Find-PSResource to use the name of the secret and the value of the secret as username/password. This wasn't the case.

* Update Register-PSResourceRepository.md - Adhere to contrib guide

Improve contrib guide adherance

* Update Register-PSResourceRepository.md

Edit for style

---------

Co-authored-by: Sean Wheeler <sean.wheeler@microsoft.com>